Move useEditLocation to PlaceStep
